requirements.txt
fastapi
uvicorn
streamlit
chromadb
langgraph
ollama
pymupdf
requests
sentence-transformers
tqdm
python-telegram-bot
python-dotenv
pymupdf-layout
pandas
pymupdf4llm
scikit-learn
rank_bm25
ragas
datasets
openpyxl
